Break out biggest pieces of options-for-select
