+++ src/js/common/schemaform/definitions/address.js
@@ -1,8 +1,38 @@
 import _ from 'lodash/fp';
 import { createSelector } from 'reselect';
 
-import { countries, states } from '../../utils/options-for-select';
-import { validateAddress } from '../validation';
+import { countries, states, isValidUSZipCode, isValidCanPostalCode } from '../../utils/address';
+
+function validatePostalCodes(errors, address) {
+  let isValidPostalCode = true;
+
+  // Checks if postal code is valid
+  if (address.country === 'USA') {
+    isValidPostalCode = isValidPostalCode && isValidUSZipCode(address.postalCode);
+  }
+  if (address.country === 'CAN') {
+    isValidPostalCode = isValidPostalCode && isValidCanPostalCode(address.postalCode);
+  }
+
+  // Add error message for postal code if it is invalid
+  if (address.postalCode && !isValidPostalCode) {
+    errors.postalCode.addError('Please provide a valid postal code');
+  }
+}
+
+const stateRequiredCountries = new Set(['USA', 'CAN', 'MEX']);
+
+function validateAddress(errors, address, formData, currentSchema) {
+  // Adds error message for state if it is blank and one of the following countries:
+  // USA, Canada, or Mexico
+  if (stateRequiredCountries.has(address.country)
+    && address.state === undefined
+    && currentSchema.required.length) {
+    errors.state.addError('Please select a state or province');
+  }
+
+  validatePostalCodes(errors, address);
+}
 
 const countryValues = countries.map(object => object.value);
 const countryNames = countries.map(object => object.label);
