+++ src/js/common/schemaform/validation.js
@@ -9,8 +9,6 @@ import {
   isValidCurrentOrFutureMonthYear,
   isValidDateRange,
   isValidRoutingNumber,
-  isValidUSZipCode,
-  isValidCanPostalCode,
   isValidPartialMonthYear,
   isValidPartialMonthYearInPast
 } from '../utils/validations';
@@ -281,36 +279,6 @@ export function validateCurrentOrPastYear(errors, year) {
   if (!isValidCurrentOrPastYear(year)) {
     errors.addError('Please provide a valid year');
   }
-}
-
-export function validatePostalCodes(errors, address) {
-  let isValidPostalCode = true;
-
-  // Checks if postal code is valid
-  if (address.country === 'USA') {
-    isValidPostalCode = isValidPostalCode && isValidUSZipCode(address.postalCode);
-  }
-  if (address.country === 'CAN') {
-    isValidPostalCode = isValidPostalCode && isValidCanPostalCode(address.postalCode);
-  }
-
-  // Add error message for postal code if it is invalid
-  if (address.postalCode && !isValidPostalCode) {
-    errors.postalCode.addError('Please provide a valid postal code');
-  }
-}
-
-const stateRequiredCountries = new Set(['USA', 'CAN', 'MEX']);
-export function validateAddress(errors, address, formData, schema) {
-  // Adds error message for state if it is blank and one of the following countries:
-  // USA, Canada, or Mexico
-  if (stateRequiredCountries.has(address.country)
-    && address.state === undefined
-    && schema.required.length) {
-    errors.state.addError('Please select a state or province');
-  }
-
-  validatePostalCodes(errors, address);
 }
 
 export function validateMatch(field1, field2) {
