+++ src/js/letters/utils/validations.js
@@ -1,4 +1,4 @@
-import { isValidUSZipCode } from '../../common/utils/validations';
+import { isValidUSZipCode } from '../../common/utils/address';
 import { ADDRESS_TYPES, MILITARY_CITIES } from './constants';
 
 const requiredMessage = 'Please enter a valid address'; // Change me!
